+++ confu/generator.py
@@ -10,12 +10,10 @@ class ConfigGenerator(object):
         pass
 
     def generate(self, schema):
-        if isclass(schema) and issubclass(schema, Schema):
+        if isinstance(schema, Schema):
             config = {}
-            for name in dir(schema):
-                attribute = getattr(schema,name)
-                if isinstance(attribute, Attribute) or (isclass(attribute) and issubclass(attribute, Schema)):
-                    config[name] = self.generate(attribute)
+            for name, attribute in schema.attributes():
+                config[name] = self.generate(attribute)
             return config
         elif isinstance(schema, Attribute):
             return self.generate(schema.default)