+++ tests/schemas.py
@@ -17,14 +17,14 @@ class NestedSchema_02(Schema):
 class Schema_01(Schema):
     int_attr = Int("int_attr")
     str_attr = Str("str_attr")
-    list_attr = List("list_attr", NestedSchema_01)
-    nested = NestedSchema_01
+    list_attr = List("list_attr", NestedSchema_01())
+    nested = NestedSchema_01("nested")
 
 class Schema_02(Schema):
     int_attr = Int("int_attr", default=123)
     str_attr = Str("str_attr", default="test")
-    list_attr = List("list_attr", NestedSchema_01)
-    nested = NestedSchema_01
+    list_attr = List("list_attr", NestedSchema_01())
+    nested = NestedSchema_01("nested")
 
 class Schema_03(Schema):
     int_attr = Int("int_attr", default=123, help="an integer attribute")
@@ -35,6 +35,6 @@ class Schema_03(Schema):
     float_attr = Float("float_attr", default=1.23, help="a float attribute")
     list_attr_int = List("list_attr_int", Int("list_attr_int_item"))
     list_attr_str = List("list_attr_str", Str("list_attr_str_item"))
-    list_attr_schema = List("list_attr_schema", NestedSchema_01)
-    nested = NestedSchema_02
+    list_attr_schema = List("list_attr_schema", NestedSchema_01())
+    nested = NestedSchema_02("nested")
     int_attr_disabled = Int("int_attr_disabled", cli=False)