+++ tests/test_generators.py
@@ -5,5 +5,5 @@ from .schemas import Schema_02
 
 def test_generate_config():
     generator = ConfigGenerator()
-    config = generator.generate(Schema_02)
+    config = generator.generate(Schema_02())
     assert config == json.loads('{"int_attr": 123, "nested": {"int_attr": null}, "list_attr": [], "str_attr": "test"}')